denti_scan_ia/API_Conecction-Joe/requirements.txt
streamlit
requests
pillow
pymongo
